+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### [Unreleased]
+## [Unreleased]
+
+## [0.14.2] - 2020-10-26
+- Install React-native-markdown-renderer
+- Fixed issue for inline text image on android devices
 
 ## [0.14.1] - 2020-10-27
 - Introduce Localizations (French)